+++ next.config.js
@@ -12,11 +12,6 @@ const nextConfig = {
       use: ["@svgr/webpack"],
     });
 
-    config.experiments = {
-      asyncWebAssembly: true,
-      layers: true,
-    };
-
     return config;
   },
   images: {